Write a NumPy program using methods info, add, array, all, greater, greater_equal, less and less_equal, equal, allclose, zeros, ones, linspace, tolist. a. To get help on the add function . b. To test whether none of the elements of a given array is zero. c. To create an element-wise comparison (greater, greater_equal, less and less_equal, equal, equal within a tolerance) of two given arrays.
Write a NumPy program using NumPy methods max, min, argmax, argmin, argmax, repr, count, bincount, unique. 
a. To extract all numbers from a given array which are less and greater than a specified number. b. To find the indices of the maximum and minimum numbers along the given axis of an array. 

Write a NumPy program using NumPy methods max, min, argmax, argmin, argmax, repr, count, bincount, unique. 
a. To extract all numbers from a given array which are less and greater than a specified number. b. To find the indices of the maximum and minimum numbers along the given axis of an array.